[][openwrt][common][bsp][Suppress Coverity warning on regs]
[Description]
Fix Coverity MISRA C-2012 Standard Libraries warning
Change input param type to unsigned for tolower()

[Release-log]
N/A

Reviewed-on: https://gerrit.mediatek.inc/c/openwrt/feeds/mtk_openwrt_feeds/+/10757199

+++ feed/app/regs/src/regs.c
@@ -171,7 +171,7 @@ int main(int argc, char **argv)
 		exit(1);
 	}
 
-	access_type = tolower(argv[1][0]);
+	access_type = tolower((unsigned int)argv[1][0]);
 	if ((access_type == 'w' && argc < 4) || (access_type == 'm' && argc < 6)) {
 		usage();
 		exit(1);
